+++ aarch32-cpu/src/register/mod.rs
@@ -27,7 +27,7 @@ pub mod dccsw;
 pub mod dcimvac;
 pub mod dcisw;
 pub mod dfar;
-#[cfg(all(target_arch = "arm", not(arm_architecture = "v4t")))]
+#[cfg(armv5te_or_higher)]
 pub mod dfsr;
 pub mod dlr;
 pub mod dracr;
@@ -53,7 +53,7 @@ pub mod id_mmfr4;
 pub mod id_pfr0;
 pub mod id_pfr1;
 pub mod ifar;
-#[cfg(all(target_arch = "arm", not(arm_architecture = "v4t")))]
+#[cfg(armv5te_or_higher)]
 pub mod ifsr;
 pub mod imp;
 pub mod iracr;
@@ -131,7 +131,7 @@ pub use dccsw::Dccsw;
 pub use dcimvac::Dcimvac;
 pub use dcisw::Dcisw;
 pub use dfar::Dfar;
-#[cfg(all(target_arch = "arm", not(arm_architecture = "v4t")))]
+#[cfg(armv5te_or_higher)]
 pub use dfsr::Dfsr;
 pub use dlr::Dlr;
 pub use dracr::Dracr;
@@ -157,7 +157,7 @@ pub use id_mmfr4::IdMmfr4;
 pub use id_pfr0::IdPfr0;
 pub use id_pfr1::IdPfr1;
 pub use ifar::Ifar;
-#[cfg(all(target_arch = "arm", not(arm_architecture = "v4t")))]
+#[cfg(armv5te_or_higher)]
 pub use ifsr::Ifsr;
 pub use iracr::Iracr;
 pub use irbar::Irbar;
@@ -250,14 +250,7 @@ pub trait SysRegRead: SysReg {
     /// side-effects that can cause Undefined Behaviour, so this method
     /// is safe.
     #[cfg_attr(not(feature = "check-asm"), inline)]
-    #[cfg_attr(
-        any(
-            arm_architecture = "v4t",
-            arm_architecture = "v5te",
-            arm_architecture = "v6"
-        ),
-        instruction_set(arm::a32)
-    )]
+    #[cfg_attr(armv6_or_lower, instruction_set(arm::a32))]
     fn read_raw() -> u32 {
         let r: u32;
         #[cfg(target_arch = "arm")]
@@ -290,14 +283,7 @@ pub trait SysRegWrite: SysReg {
     /// You need to read the Architecture Reference Manual to verify that you are
     /// writing valid data here.
     #[cfg_attr(not(feature = "check-asm"), inline)]
-    #[cfg_attr(
-        any(
-            arm_architecture = "v4t",
-            arm_architecture = "v5te",
-            arm_architecture = "v6"
-        ),
-        instruction_set(arm::a32)
-    )]
+    #[cfg_attr(armv6_or_lower, instruction_set(arm::a32))]
     unsafe fn write_raw(_value: u32) {
         #[cfg(target_arch = "arm")]
         unsafe {