+++ examples/mps3-an536/src/bin/gic-map.rs
@@ -77,11 +77,9 @@ fn main() -> ! {
     });
 
     println!("Enabling interrupts...");
-    dump_sctlr();
     unsafe {
         aarch32_cpu::interrupt::enable();
     }
-    dump_sctlr();
 
     // Send it
     println!("Send lo-prio SGI");
@@ -104,11 +102,6 @@ fn main() -> ! {
     println!("IRQ test completed OK");
 
     mps3_an536::exit(0);
-}
-
-fn dump_sctlr() {
-    let sctlr = aarch32_cpu::register::Sctlr::read();
-    println!("{:?}", sctlr);
 }
 
 /// Handles the low-prio SGI