+++ examples/mps3-an536/src/bin/registers.rs
@@ -36,7 +36,7 @@ fn chip_info() {
 #[cfg(arm_architecture = "v7-r")]
 fn mpu_pmsa_v7() {
     use aarch32_cpu::{
-        pmsav7::{Cacheable, Config, MemAttr, Mpu, Region, RegionSize},
+        pmsav7::{AccessPerms, Cacheable, Config, MemAttr, Mpu, Region, RegionSize},
         register::Mpuir,
     };
 
@@ -73,6 +73,7 @@ fn mpu_pmsa_v7() {
                 outer: Cacheable::NonCacheable,
                 shareable: true,
             },
+            access_perms: AccessPerms::ReadWrite,
         }],
         iregions: &[],
     })